+++ src/views/dashboard/related/topology/components/utils/layout.ts
@@ -120,3 +120,115 @@ export function circleIntersection(ax: number, ay: number, ar: number, bx: numbe
     { x: gx, y: gy },
   ];
 }
+function findMostFrequent(arr: Call[]) {
+  const count: any = {};
+  let maxCount = 0;
+  let maxItem = null;
+
+  for (let i = 0; i < arr.length; i++) {
+    const item = arr[i];
+    count[item.sourceObj.id] = (count[item.sourceObj.id] || 0) + 1;
+    if (count[item.sourceObj.id] > maxCount) {
+      maxCount = count[item.sourceObj.id];
+      maxItem = item.sourceObj;
+    }
+    count[item.targetObj.id] = (count[item.targetObj.id] || 0) + 1;
+    if (count[item.targetObj.id] > maxCount) {
+      maxCount = count[item.targetObj.id];
+      maxItem = item.targetObj;
+    }
+  }
+
+  return maxItem;
+}
+export function computeLevels(calls: Call[], nodeList: Node[], levels: any[]) {
+  const node = findMostFrequent(calls);
+  const nodes = JSON.parse(JSON.stringify(nodeList)).sort((a: Node, b: Node) => {
+    if (a.name.toLowerCase() < b.name.toLowerCase()) {
+      return -1;
+    }
+    if (a.name.toLowerCase() > b.name.toLowerCase()) {
+      return 1;
+    }
+    return 0;
+  });
+  const index = nodes.findIndex((n: Node) => n.type === "USER");
+  let key = index;
+  if (index < 0) {
+    key = nodes.findIndex((n: Node) => n.id === node.id);
+  }
+  levels.push([nodes[key]]);
+  nodes.splice(key, 1);
+  for (const level of levels) {
+    const a = [];
+    for (const l of level) {
+      for (const n of calls) {
+        if (n.target === l.id) {
+          const i = nodes.findIndex((d: Node) => d.id === n.source);
+          if (i > -1) {
+            a.push(nodes[i]);
+            nodes.splice(i, 1);
+          }
+        }
+        if (n.source === l.id) {
+          const i = nodes.findIndex((d: Node) => d.id === n.target);
+          if (i > -1) {
+            a.push(nodes[i]);
+            nodes.splice(i, 1);
+          }
+        }
+      }
+    }
+    if (a.length) {
+      levels.push(a);
+    }
+  }
+  if (nodes.length) {
+    const ids = nodes.map((d: Node) => d.id);
+    const links = calls.filter((item: Call) => ids.includes(item.source) || ids.includes(item.target));
+    const list = computeLevels(links, nodes, []);
+    levels = list.map((subArrayA, index) => subArrayA.concat(levels[index]));
+  }
+  return levels;
+}
+export function changeNode(
+  d: { x: number; y: number },
+  currentNode: Nullable<Node>,
+  topologyLayout: any,
+  radius: number,
+) {
+  if (!currentNode) {
+    return;
+  }
+  for (const node of topologyLayout.nodes) {
+    if (node.id === currentNode.id) {
+      node.x = d.x;
+      node.y = d.y;
+    }
+  }
+  for (const call of topologyLayout.calls) {
+    if (call.sourceObj.id === currentNode.id) {
+      call.sourceObj.x = d.x;
+      call.sourceObj.y = d.y;
+    }
+    if (call.targetObj.id === currentNode.id) {
+      call.targetObj.x = d.x;
+      call.targetObj.y = d.y;
+    }
+    if (call.targetObj.id === currentNode.id || call.sourceObj.id === currentNode.id) {
+      const pos: any = circleIntersection(
+        call.sourceObj.x,
+        call.sourceObj.y,
+        radius,
+        call.targetObj.x,
+        call.targetObj.y,
+        radius,
+      );
+      call.sourceX = pos[0].x;
+      call.sourceY = pos[0].y;
+      call.targetX = pos[1].x;
+      call.targetY = pos[1].y;
+    }
+  }
+  return computeCallPos(topologyLayout.value.calls, radius);
+}
